Merge

This family of functions revolves around merging the intervals within a single iv.

iv_merge() merges overlapping or abutting intervals in x.

iv_replace_merged() replaces each interval in x with the merged interval that it maps to. This is particularly useful alongside dplyr::group_by().

iv_locate_merge_bounds() returns locations for slicing the iv_start() and iv_end() of x to generate the merged result.

iv_locate_merge_groups() returns a two column data frame with a key column containing the result of iv_locate_merge_bounds() and a loc list-column containing integer vectors that map each element of x to the merged interval that it falls in.

Optionally, you can choose not to merge abutting intervals with abutting = FALSE, which can be useful if you'd like to retain those boundaries.

Minimal interval vectors

iv_merge() is particularly useful because it can generate a minimal interval vector, which covers the range of an interval vector in the most compact form possible. In particular, a minimal interval vector:

Has no overlapping intervals

Has no abutting intervals

Is ordered on both start and end

A minimal interval vector is allowed to have a single missing interval, which is located at the end of the vector.

Usage

iv_merge(x, ..., abutting = TRUE)

iv_replace_merged(x, ..., abutting = TRUE)

iv_locate_merge_bounds(x, ..., abutting = TRUE)

iv_locate_merge_groups(x, ..., abutting = TRUE)

Arguments

x

[iv]

An interval vector.

...

These dots are for future extensions and must be empty.

abutting

[logical(1)]

Should abutting intervals be merged?

If TRUE, [a, b) and [b, c) will merge as [a, c). If FALSE, they will be kept separate. To be a minimal interval vector, all abutting intervals must be merged.

Value

For iv_merge(), an iv with the same type as x.

For iv_replace_merged(), an iv with the same type and size as x.

For iv_locate_merge_bounds(), a two column data frame with start and end integer columns.

For iv_locate_merge_groups(), a two column data frame with a key column containing the result of iv_locate_merge_bounds() and a loc list-column containing integer vectors.

Examples

library(dplyr, warn.conflicts = FALSE)

x <- iv_pairs(
  c(1, 5),
  c(2, 3),
  c(NA, NA),
  c(5, 6),
  c(NA, NA),
  c(9, 12),
  c(11, 14)
)
x
#> <iv<double>[7]>
#> [1] [1, 5)   [2, 3)   [NA, NA) [5, 6)   [NA, NA) [9, 12)  [11, 14)

# Merging removes all redundancy while still covering the full range
# of values that were originally represented. If any missing intervals
# are present, a single one is retained.
iv_merge(x)
#> <iv<double>[3]>
#> [1] [1, 6)   [9, 14)  [NA, NA)

# Abutting intervals are typically merged, but you can choose not to merge
# them if you want to retain those boundaries
iv_merge(x, abutting = FALSE)
#> <iv<double>[4]>
#> [1] [1, 5)   [5, 6)   [9, 14)  [NA, NA)

# `iv_replace_merged()` is useful alongside `group_by()` and `summarize()`
df <- tibble(x = x)
df <- mutate(df, u = iv_replace_merged(x))
df
#> # A tibble: 7 × 2
#>           x         u
#>   <iv<dbl>> <iv<dbl>>
#> 1    [1, 5)    [1, 6)
#> 2    [2, 3)    [1, 6)
#> 3  [NA, NA)  [NA, NA)
#> 4    [5, 6)    [1, 6)
#> 5  [NA, NA)  [NA, NA)
#> 6   [9, 12)   [9, 14)
#> 7  [11, 14)   [9, 14)

df %>%
  group_by(u) %>%
  summarize(n = n())
#> # A tibble: 3 × 2
#>           u     n
#>   <iv<dbl>> <int>
#> 1    [1, 6)     3
#> 2   [9, 14)     2
#> 3  [NA, NA)     2

# The real workhorse here is `iv_locate_merge_groups()`, which returns
# information on where to slice `x` to get the merged result, and which
# observations of `x` belong to which merged interval
iv_locate_merge_groups(x)
#>   key.start key.end     loc
#> 1         1       4 1, 2, 4
#> 2         6       7    6, 7
#> 3         3       5    3, 5