Splits

This family of functions revolves around splitting an iv on its endpoints, which results in a new iv that is disjoint (i.e. non-overlapping).

iv_split() splits x into a new iv that is completely disjoint.

iv_replace_splits() replaces x with a list of the same size where each element of the list contains the disjoint iv resulting from splitting that element of x. This is particularly useful alongside tidyr::unnest().

iv_locate_split_groups() returns a two column data frame with a key column containing the result of iv_split() and a loc list-column containing integer vectors that map each element of x to the disjoint intervals that it falls in.

iv_split(x)

iv_replace_splits(x)

iv_locate_split_groups(x)

Arguments

x

[iv]

An interval vector.

Value

For iv_split(), an iv with the same type as x. For iv_replace_splits(), a list-of containing ivs with the same size as x. For iv_locate_split_groups(), a two column data frame with a key column of the same type as x and loc list-column containing integer vectors.

Examples

library(tidyr)
library(dplyr)

# Guests to a party and their arrival/departure times
guests <- tibble(
  arrive = as.POSIXct(
    c("2008-05-20 19:30:00", "2008-05-20 20:10:00", "2008-05-20 22:15:00"),
    tz = "UTC"
  ),
  depart = as.POSIXct(
    c("2008-05-20 23:00:00", "2008-05-21 00:00:00", "2008-05-21 00:30:00"),
    tz = "UTC"
  ),
  name = list(
    c("Mary", "Harry"),
    c("Diana", "Susan"),
    "Peter"
  )
)

guests <- unnest(guests, name) %>%
  mutate(iv = iv(arrive, depart), .keep = "unused")

guests
#> # A tibble: 5 × 2
#>   name                                          iv
#>   <chr>                                 <iv<dttm>>
#> 1 Mary  [2008-05-20 19:30:00, 2008-05-20 23:00:00)
#> 2 Harry [2008-05-20 19:30:00, 2008-05-20 23:00:00)
#> 3 Diana [2008-05-20 20:10:00, 2008-05-21 00:00:00)
#> 4 Susan [2008-05-20 20:10:00, 2008-05-21 00:00:00)
#> 5 Peter [2008-05-20 22:15:00, 2008-05-21 00:30:00)

# You can determine the disjoint intervals at which people
# arrived/departed with `iv_split()`
iv_split(guests$iv)
#> <iv<datetime<UTC>>[5]>
#> [1] [2008-05-20 19:30:00, 2008-05-20 20:10:00)
#> [2] [2008-05-20 20:10:00, 2008-05-20 22:15:00)
#> [3] [2008-05-20 22:15:00, 2008-05-20 23:00:00)
#> [4] [2008-05-20 23:00:00, 2008-05-21 00:00:00)
#> [5] [2008-05-21 00:00:00, 2008-05-21 00:30:00)

# Say you'd like to determine who was at the party at any given time
# throughout the night
guests <- mutate(guests, splits = iv_replace_splits(iv))
guests
#> # A tibble: 5 × 3
#>   name                                          iv           splits
#>   <chr>                                 <iv<dttm>> <list<iv<dttm>>>
#> 1 Mary  [2008-05-20 19:30:00, 2008-05-20 23:00:00)              [3]
#> 2 Harry [2008-05-20 19:30:00, 2008-05-20 23:00:00)              [3]
#> 3 Diana [2008-05-20 20:10:00, 2008-05-21 00:00:00)              [3]
#> 4 Susan [2008-05-20 20:10:00, 2008-05-21 00:00:00)              [3]
#> 5 Peter [2008-05-20 22:15:00, 2008-05-21 00:30:00)              [3]

# Unnest the splits to generate disjoint intervals for each guest
guests <- guests %>%
  unnest(splits) %>%
  select(name, splits)

guests
#> # A tibble: 15 × 2
#>    name                                      splits
#>    <chr>                                 <iv<dttm>>
#>  1 Mary  [2008-05-20 19:30:00, 2008-05-20 20:10:00)
#>  2 Mary  [2008-05-20 20:10:00, 2008-05-20 22:15:00)
#>  3 Mary  [2008-05-20 22:15:00, 2008-05-20 23:00:00)
#>  4 Harry [2008-05-20 19:30:00, 2008-05-20 20:10:00)
#>  5 Harry [2008-05-20 20:10:00, 2008-05-20 22:15:00)
#>  6 Harry [2008-05-20 22:15:00, 2008-05-20 23:00:00)
#>  7 Diana [2008-05-20 20:10:00, 2008-05-20 22:15:00)
#>  8 Diana [2008-05-20 22:15:00, 2008-05-20 23:00:00)
#>  9 Diana [2008-05-20 23:00:00, 2008-05-21 00:00:00)
#> 10 Susan [2008-05-20 20:10:00, 2008-05-20 22:15:00)
#> 11 Susan [2008-05-20 22:15:00, 2008-05-20 23:00:00)
#> 12 Susan [2008-05-20 23:00:00, 2008-05-21 00:00:00)
#> 13 Peter [2008-05-20 22:15:00, 2008-05-20 23:00:00)
#> 14 Peter [2008-05-20 23:00:00, 2008-05-21 00:00:00)
#> 15 Peter [2008-05-21 00:00:00, 2008-05-21 00:30:00)

# Tabulate who was there at any given time
guests %>%
  group_by(splits) %>%
  summarise(n = n(), who = list(name))
#> # A tibble: 5 × 3
#>                                       splits     n who      
#>                                   <iv<dttm>> <int> <list>   
#> 1 [2008-05-20 19:30:00, 2008-05-20 20:10:00)     2 <chr [2]>
#> 2 [2008-05-20 20:10:00, 2008-05-20 22:15:00)     4 <chr [4]>
#> 3 [2008-05-20 22:15:00, 2008-05-20 23:00:00)     5 <chr [5]>
#> 4 [2008-05-20 23:00:00, 2008-05-21 00:00:00)     3 <chr [3]>
#> 5 [2008-05-21 00:00:00, 2008-05-21 00:30:00)     1 <chr [1]>