Union

This family of functions revolves around taking the union of intervals within a single interval vector.

iv_union() returns the union of all the intervals in x.

iv_replace_union() replaces each value in x with its union interval. This is particularly useful alongside dplyr::group_by().

iv_locate_union() returns locations for slicing the iv_start() and iv_end() of x to generate the union.

iv_locate_union_groups() returns the same locations as iv_locate_union() along with a list-column of group locations that maps each element of x to its union interval.

Optionally, you can choose to keep abutting intervals separate with keep_abutting, which can be useful if you'd like to retain those boundaries.

Minimal interval vectors

iv_union() is particularly useful because it can generate a minimal interval vector, which covers the range of an interval vector in the most compact form possible. In particular, a minimal interval vector:

Has no overlapping intervals

Has no abutting intervals

Is ordered on both start and end

A minimal interval vector is allowed to have a single missing interval, which is located at the end of the vector.

iv_union(x, ..., keep_abutting = FALSE)

iv_replace_union(x, ..., keep_abutting = FALSE)

iv_locate_union(x, ..., keep_abutting = FALSE)

iv_locate_union_groups(x, ..., keep_abutting = FALSE)

Arguments

x

[iv]

An interval vector.

...

These dots are for future extensions and must be empty.

keep_abutting

[logical(1)]

Should abutting intervals be kept separate?

If FALSE, [a, b) and [b, c) will join as [a, c). If TRUE, they will be kept separate. To be a minimal interval vector, all abutting intervals must be merged.

Examples

library(dplyr, warn.conflicts = FALSE)

x <- iv_pairs(
  c(1, 5),
  c(2, 3),
  c(NA, NA),
  c(5, 6),
  c(NA, NA),
  c(9, 12),
  c(11, 14)
)
x
#> <iv<double>[7]>
#> [1] [1, 5)   [2, 3)   [NA, NA) [5, 6)   [NA, NA) [9, 12)  [11, 14)

# The union removes all redundancy while still covering the full range
# of values that were originally represented. If any missing intervals
# are present, a single one is retained.
iv_union(x)
#> <iv<double>[3]>
#> [1] [1, 6)   [9, 14)  [NA, NA)

# Abutting intervals can be kept separate if you want to
# retain those boundaries
iv_union(x, keep_abutting = TRUE)
#> <iv<double>[4]>
#> [1] [1, 5)   [5, 6)   [9, 14)  [NA, NA)

# `iv_replace_union()` is useful alongside `group_by()` and `summarize()`
df <- tibble(x = x)
df <- mutate(df, u = iv_replace_union(x))
df
#> # A tibble: 7 × 2
#>           x         u
#>   <iv<dbl>> <iv<dbl>>
#> 1    [1, 5)    [1, 6)
#> 2    [2, 3)    [1, 6)
#> 3  [NA, NA)  [NA, NA)
#> 4    [5, 6)    [1, 6)
#> 5  [NA, NA)  [NA, NA)
#> 6   [9, 12)   [9, 14)
#> 7  [11, 14)   [9, 14)

df %>%
  group_by(u) %>%
  summarize(n = n())
#> # A tibble: 3 × 2
#>           u     n
#>   <iv<dbl>> <int>
#> 1    [1, 6)     3
#> 2   [9, 14)     2
#> 3  [NA, NA)     2

# The real workhorse here is `iv_locate_union_groups()`, which returns
# information on where to slice `x` to get the union, and which observations
# of `x` belong to which union interval
iv_locate_union_groups(x)
#>   key.start key.end     loc
#> 1         1       4 1, 2, 4
#> 2         6       7    6, 7
#> 3        NA      NA    3, 5