Complement

iv_complement() computes the set theoretical complement of an interval vector. The "universe" to compute the complement over is considered to be the minimal and maximal values of x, but this can be adjusted with lower and upper.

iv_complement(x, ..., lower = NULL, upper = NULL)

Arguments

x

[iv]

An interval vector.

...

These dots are for future extensions and must be empty.

lower, upper

[vector(1) / NULL]

Bounds for the universe over which to compute the complement. These should have the same type as the element type of the interval vector. It is often useful to expand the universe to, say, -Inf to Inf.

Value

An object with the same type as x representing its complement.

Examples

x <- new_iv(c(10, 0, 3, -5), c(12, 5, 6, -2))
x
#> <iv[4]>
#> [1] [10, 12) [0, 5)   [3, 6)   [-5, -2)

# Complement contains any values from `[-5, 12)` that aren't represented
# in these intervals
iv_complement(x)
#> <iv[2]>
#> [1] [-2, 0) [6, 10)

# Expand out the "universe" of possible values
iv_complement(x, lower = -Inf)
#> <iv[3]>
#> [1] [-Inf, -5) [-2, 0)    [6, 10)   
iv_complement(x, lower = -Inf, upper = Inf)
#> <iv[4]>
#> [1] [-Inf, -5) [-2, 0)    [6, 10)    [12, Inf)