Detect a relationship in parallel between two ivs

This family of functions detects different types of relationships between two ivs in parallel, where parallel means that the i-th interval of x is compared against the i-th interval of y. This is in contrast to iv_detect_overlaps(), which works more like base::%in%.

iv_detect_parallel_overlaps() detects a specific type of overlap between the i-th interval of x and the i-th interval of y.

iv_detect_parallel_precedes() detects if the i-th interval of x precedes (i.e. comes before) the i-th interval of y.

iv_detect_parallel_follows() detects if the i-th interval of x follows (i.e. comes after) the i-th interval of y.

These functions return a logical vector the same size as the common size of x and y.

Missing intervals

Missing intervals in x or y force an error to be thrown by default, as it is unclear what the desired result is when missing intervals are involved. This is consistent with iv_detect_overlaps(). If you'd like to match missing intervals exactly, set missing = "match". If you'd like missing intervals to be unmatched, set missing = FALSE. If you'd like missing intervals to be propagated, set missing = NA.

iv_detect_parallel_overlaps(x, y, ..., type = "any", missing = "error")

iv_detect_parallel_precedes(x, y, ..., missing = "error")

iv_detect_parallel_follows(x, y, ..., missing = "error")

Arguments

x, y

[iv]

A pair of interval vectors.

These will be recycled against each other and cast to the same type.

...

These dots are for future extensions and must be empty.

type

[character(1)]

The type of relationship to find. One of:

"any": Finds any overlap whatsoever between an interval in needles and an interval in haystack.

"within": Finds when an interval in needles is completely within (or equal to) an interval in haystack.

"contains": Finds when an interval in needles completely contains (or equals) an interval in haystack.

"equals": Finds when an interval in needles is exactly equal to an interval in haystack.

"starts": Finds when the start of an interval in needles matches the start of an interval in haystack.

"ends": Finds when the end of an interval in needles matches the end of an interval in haystack.

missing

[logical(1) / "match" / "error"]

Handling of missing intervals in x or y.

"match" matches missing intervals in x to missing intervals in y. Regardless of type, if both intervals are missing, then this results in TRUE, otherwise it results in FALSE.

"error" throws an error if any intervals are missing. This is the default.

If a single logical value is provided, this represents the value returned in the i-th element of the result if missing intervals are present in the i-th interval of x or y. You can force missing intervals to be unmatched by setting this to FALSE, and you can force them to be propagated by setting this to NA.

Value

A logical vector the same size as the common size of x and y.

See also

Examples

library(vctrs)

x <- iv_pairs(
  as.Date(c("2019-01-05", "2019-01-10")),
  as.Date(c("2019-01-07", "2019-01-15")),
  as.Date(c("2019-01-20", "2019-01-31"))
)

y <- iv_pairs(
  as.Date(c("2019-01-01", "2019-01-03")),
  as.Date(c("2019-01-07", "2019-01-09")),
  as.Date(c("2019-01-18", "2019-01-21"))
)

x
#> <iv<date>[3]>
#> [1] [2019-01-05, 2019-01-10) [2019-01-07, 2019-01-15) [2019-01-20, 2019-01-31)
y
#> <iv<date>[3]>
#> [1] [2019-01-01, 2019-01-03) [2019-01-07, 2019-01-09) [2019-01-18, 2019-01-21)

# Does the i-th interval of `x` overlap the i-th interval of `y`?
iv_detect_parallel_overlaps(x, y)
#> [1] FALSE  TRUE  TRUE

# Does the i-th interval of `x` contain the i-th interval of `y`?
iv_detect_parallel_overlaps(x, y, type = "contains")
#> [1] FALSE  TRUE FALSE

# Does the i-th interval of `x` follow the i-th interval of `y`?
iv_detect_parallel_follows(x, y)
#> [1]  TRUE FALSE FALSE

# ---------------------------------------------------------------------------

a <- iv(c(1, NA), c(2, NA))
b <- iv(c(NA, NA), c(NA, NA))

# Missing intervals error by default
try(iv_detect_parallel_overlaps(a, b))
#> Error in iv_detect_parallel_overlaps(a, b) : 
#>   Can't have missing values in `x` or `y`.
#> ℹ A value at location 1 is missing.
#> ℹ Use `missing` to control how missing values should be handled if they are expected.

# If you'd like missing intervals to match exactly, set `missing = "match"`
iv_detect_parallel_overlaps(a, b, missing = "match")
#> [1] FALSE  TRUE

# If you'd like missing intervals to be treated as unmatched, set
# `missing = FALSE`
iv_detect_parallel_overlaps(a, b, missing = FALSE)
#> [1] FALSE FALSE

# If you'd like to propagate missing intervals, set `missing = NA`
iv_detect_parallel_overlaps(a, b, missing = NA)
#> [1] NA NA